(feat): basic TCP proxy
Implemented a very basic async TCP proxy using tokio that accepts any
number of connection to a specified listener and streams bidirectionally
between the client and target. Listener address and port and target
address and port are specified in the CLI by the user; this was
implemented using the clap crate.