(refactor): socks5 implementation
Furthered the SOCKS5 implementation by building the required structs
and enums to model the SOCKS5 protocol according to the spec. I
implemented a number of functions to handle socks handshakes and
requests. I need to still implement functionality to connect to the
target and start the bidirectional streaming. I also need to implement a
send_repy function to handle replies from the server to client.