(refactor): socks5 crate
I refactored the socks5 crate to separate concerns a bit. The socks5.rs
file was quickly approaching 400 lines, and I wanted to make it a bit
more modular and readable. socks5 now lives in its own directory in src/
and is split amongst files based on function: auth, commands, protocol,
and server. The functionality is currently exactly the same.