(feat): username/pass auth
I added username/password authentication to the SOCKS5 server according
to RFC1929. This required the refactoring of the server code: creating a
server struct and new, with_auth, and run functions. Much of this was
largely just moved from main. The username/password are housed in an
UserPass struct in the auth.rs file. Ownernship of this data is shared
via std::sync:Arc. The relevant functions were added to auth to handle
the authentication itself.