(feat): udp associate command
Began implementation of the UDP associate command for the SOCKS5 proxy.
This included the creation of two a TransportProcol enum and
UdpAssociate struct in the commands mod. I ended up needing to change
some function signatures in the commands and server mods in order to
return values/objects that could handle both tcp and udp connections as
well as moving some of the logic from server to a handle_socks_request
function in commands.