(refactor): server.rs
I refactored the server.rs mod to make it more readable and in line with
actual server functionality. SOCKS5 command-related functionality such
as the relaying TCP or UDP was bleeding in, and it seemed to make more
sense in the commands.rs mod. I created a Connect struct and made a run
method to handle relaying. I added an implementation block and the
skeleton for a run function for UdpAssociate. I also added some
constants in protocol that will prove useful when parsing and rebuilding
UDP headers during UdpAssociate relaying.