(fix): authentication logic bug
There was a bug in the authentication logic when setting the server to
require username/password authentication. The bug was in the
select_auth_method function, which was structured to select an
authentication method from a array of methods in order of preference. It
would compare these to those provided by the client and accept the
"best" one. However, when a user connected, it would select NoAuth and
not require username/password authentication from the user.

The new logic checks for whether authentication configuration has been
provided via a new parameter to the function and operates on the logic
accordingly to only allow username/password authentication when it has
been provided on the server side.