(refactor): commands.rs cleanup
Removed some dead code from command.rs, namely the socket_to_association
field in the UdpRelayState struct. I realized that I was instantiating
the field and removing it during cleanup but was using it to determine
the client address when relaying traffic back from the target. I am
using the outbound_sockets field, which already contains that data. The
client_addr mapping we are interested in is initially retrieved during
the reception from a client's datagram in the call to recv_from and then
passed through the entire chain -> handle_client_datagram,
monitor_outbound_socket, then sent through the channel, which is
received in the select loop and passed to send_response_to_client.