(refactor): UdpAssociate struct cleanup
I removed the target_addr from the UdpAssociate struct as it doesn't
need to be tracked there. Up to this point, I had been parsing the
target_addr from the initial UDP ASSOCIATE request and storing it in a
UdpAssociate instantiation. However, I realized this was unnecessary as
I am parsing the target address from each client datagram in
handle_client_datagram. According to to RFC 1928, the client *can*
provide the address and port that it expects to use, which the server
*may* use to limit access to the association. However, in practice,
clients typically send a value of 0.0.0.0:0, which is required by the
protocol if a destination is unknown at the time of association.

I removed the field from the struct and simplified the
handle_udpassociate_cmd function to ignore the initial address value.