(feat): add Socks5Server::bind method
This is also a breaking change as it makes run mutable in order to
reuse code.

Alternative implementations could use a type-state to enforce bind
is not called twice, or listener is accepted as an input parameter,
or there is some code duplication for the logging to not need the
mutation.

The log has also changed to show the actual local_addr instead of
the input provided by the user, since these can be different.